products section on process

+++ model/src/main/java/com/android/model/repository/ingredients/IngredientsRepository.kt
@@ -47,7 +47,7 @@ class IngredientsRepository @Inject constructor(
    }
 
    suspend fun addExpensesAndIncomesOfIngredient(
-      ingredientId: Int?,
+      ingredientId: Long?,
       amount: String,
       statusOfActions: Int,
       comments: String
