+++ model/src/main/java/com/android/model/repository/ingredients/IngredientsSource.kt
@@ -36,7 +36,7 @@ interface IngredientsSource {
     * @throws ParseBackendResponseException
     */
    suspend fun addExpensesAndIncomesOfIngredient(
-      ingredientId: Int,
+      ingredientId: Long,
       amount: String,
       statusOfActions: Int,
       comments: String
